requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
